Cleanup intrinsic handling
