+++ Sources/Placement/Layouting/Coordinator.swift
@@ -6,10 +6,6 @@ class PlacementsCoordinator: ObservableObject {
     var unspecifiedSize: [AnyHashable: CGSize] = [:]
 }
 
-class LayoutSizingCoordinator: ObservableObject {
-    
-}
-
 class Coordinator<L: PlacementLayout>: ObservableObject {
     var safeAreaInsets: UIEdgeInsets? = nil
     var globalFrame: CGRect? = nil
@@ -191,6 +187,5 @@ class Coordinator<L: PlacementLayout>: ObservableObject {
         
     public var transaction = Transaction()
     public var placementsCoordinator = PlacementsCoordinator()
-    public var layoutSizingCoordinator = LayoutSizingCoordinator()
     public var hostingControllers: [AnyHashable: UIHostingController<AnyView>] = [:]
 }
