+++ Sources/Placement/Layouting/Layouter.swift
@@ -29,7 +29,6 @@ struct Layouter<Content: View, L: PlacementLayout>: View {
         }
         .environmentObject(coordinator)
         .environmentObject(coordinator.placementsCoordinator)
-        .environmentObject(coordinator.layoutSizingCoordinator)
         .transformPreference(PlacementIntrinsicSizesPreferenceKey.self) { anchors in
             anchors = [:]
         }
